add forAll and forNone to all data driven test roots and remove extra file
