+++ kotest-framework/kotest-framework-engine/src/jvmTest/kotlin/com/sksamuel/kotest/engine/datatest/WordSpecDataTest.kt
@@ -1,5 +1,6 @@
 package com.sksamuel.kotest.engine.datatest
 
+import io.kotest.assertions.withClue
 import io.kotest.core.annotation.Ignored
 import io.kotest.core.spec.style.WordSpec
 import io.kotest.core.datatest.forAll
@@ -11,12 +12,15 @@ internal class WordSpecDataTest : WordSpec() {
    init {
       data class PythagTriple(val a: Int, val b: Int, val c: Int)
 
-      forAll(
-         PythagTriple(3, 4, 5),
-         PythagTriple(6, 8, 10),
-      ) { (a, b, c) ->
-         "root datatest forAll" should {
-            a * a + b * b shouldBe c * c
+      forAll(2, 4, 6) {
+         withClue("$it is even check") {
+            it % 2 shouldBe 0
+         }
+      }
+
+      forNone(1, 3, 5) {
+         withClue("$it is even check") {
+            it % 2 shouldBe 0
          }
       }
